requirements.txt
# Django Web Framework
Django==6.0.2
djangorestframework==3.16.1

# Data Science & Machine Learning
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.6.1
joblib==1.4.2

# Data Visualization
matplotlib==3.9.2
seaborn==0.13.2

# Jupyter Notebook
jupyter==1.1.1
notebook==7.3.1
